Home controller refactory
