--- code/SoloSavingsApp/src/main/java/com/soloSavings/controller/CustomErrorController.java
@@ -1,15 +1,0 @@
-package com.soloSavings.controller;
-
-import org.springframework.boot.web.servlet.error.ErrorController;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class CustomErrorController implements ErrorController {
-
-    @RequestMapping("/error")
-    public String handleError() {
-        // Handle errors and return the appropriate template
-        return "error"; // For error JSP page
-    }
-}
