+++ code/SoloSavingsApp/src/main/java/com/soloSavings/controller/HomeController.java
@@ -11,6 +11,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.boot.web.servlet.error.ErrorController;
 
 @Controller
 public class HomeController {
@@ -28,4 +30,20 @@ public class HomeController {
         return "home";
     }
 
+    @GetMapping("/solosavings/register")
+    public String register() {
+        logger.info(" Register an account with SoloSavings Application");
+        return "register";
+    }
+    @GetMapping("/solosavings/login")
+    public String login() {
+        logger.info(" Login to SoloSavings Application");
+        return "login";
+    }
+
+    @RequestMapping("/error")
+    public String handleError() {
+        // Handle errors and return the appropriate template
+        return "error"; // For error JSP page
+    }
 }
