+++ code/SoloSavingsApp/src/main/java/com/soloSavings/controller/UserAuthenticationController.java
@@ -2,8 +2,8 @@ package com.soloSavings.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 /*
  * Copyright (c) 2023 Team 2 - SoloSavings
@@ -13,17 +13,10 @@ import org.springframework.web.bind.annotation.GetMapping;
  * This software is the confidential and proprietary information of
  * Team 2 - SoloSavings Application
  */
-@Controller
+
+@RestController
+@RequestMapping("/api")
 public class UserAuthenticationController {
     private static final Logger logger = LoggerFactory.getLogger(HomeController.class);
-    @GetMapping("/solosavings/register")
-    public String register() {
-        logger.info(" Register an account with SoloSavings Application");
-        return "register";
-    }
-    @GetMapping("/solosavings/login")
-    public String login() {
-        logger.info(" Login to SoloSavings Application");
-        return "login";
-    }
+
 }
