+++ code/SoloSavingsApp/src/main/resources/application.properties
@@ -11,3 +11,5 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 server.port=8888
 server.error.path=/error
 logging.level.org.springframework.security=TRACE
+jwt.secret=mysecretkey
+jwt.expiration=86400
